BootloaderTest: Fix serial number format check
This test had some of the byte offsets incorrect, which happened to not
cause a problem with the particular G031 serial number this was tested
with.

Additionally, the serial number byte order was changed, so the checks
are adapted to that.

Finally, it seems that the "letter followed by two digits" seen in the
G031 lot number is not universal, the G030 we're now testing with has
mixed letters and digits, so change the checks accordingly.