bootloader: Increase MAX_PACKET_LENGTH to 255 on hopperboard
This helps to speed up big firmware uploads. With this change and the
current 1Mbps and 150μs inter-frame spacing, a full 64k flash upload is
reduced from just over 1.1s to 0.7s. With the default 19200bps and
1750μs spacing, this change reduces the time from 43s to 35s.

Note that the time to actually write to flash is not included in these
numbers (which is typically 1.4s for regular programming and 1.1s for
fast programming).